tools/requirements.txt
pdfplumber
argparse
pathlib
jinja2
rapidfuzz
python-docx